Correct the clear_directory function when no dir actually exist

+++ Training/dataset_exporter.py
@@ -6,18 +6,20 @@ import shutil
 def clear_directory(directory):
     """
     Clear all files and subdirectories in the specified directory.
-
     :param directory: The directory to clear.
     """
-    for filename in os.listdir(directory):
-        file_path = os.path.join(directory, filename)
-        try:
-            if os.path.isfile(file_path) or os.path.islink(file_path):
-                os.unlink(file_path)
-            elif os.path.isdir(file_path):
-                shutil.rmtree(file_path)
-        except Exception as e:
-            print(f'Failed to delete {file_path}. Reason: {e}')
+    if os.path.exists(directory):
+        for filename in os.listdir(directory):
+            file_path = os.path.join(directory, filename)
+            try:
+                if os.path.isfile(file_path) or os.path.islink(file_path):
+                    os.unlink(file_path)
+                elif os.path.isdir(file_path):
+                    shutil.rmtree(file_path)
+            except Exception as e:
+                print(f'Failed to delete {file_path}. Reason: {e}')
+    else:
+        print(f'Directory not found: {directory}')
 
 def create_directories(base_dir, sub_dirs):
     """
